Added some value initialization. Added some missing documentation.

+++ src/api/web/gw2/mapping/v2/characters/inventory/Inventory.java
@@ -25,7 +25,7 @@ public interface Inventory {
 
     /**
      * Gets the number of this object in inventory.
-     * @return An {@code int}.
+     * @return A {@code short}.
      */
-    int getCount();
+    short getCount();
 }
